+++ srcs/ft_mouse_hook.c
@@ -7,7 +7,7 @@ static double	ft_intrpl(double start, double end, double itp)
 }
 */
 
-static void ft_translation_back(t_data *data, double x_cp, double y_cp)
+static void ft_translation_back(t_data *data, long double x_cp, long double y_cp)
 {
 	data->min_r = data->min_r + x_cp;
 	data->max_r = data->max_r + x_cp;
@@ -15,7 +15,7 @@ static void ft_translation_back(t_data *data, double x_cp, double y_cp)
 	data->max_i = data->max_i + y_cp;
 }
 
-static void ft_translation_to_zero(t_data *data, double x_cp, double y_cp)
+static void ft_translation_to_zero(t_data *data, long double x_cp, long double y_cp)
 {
 	data->min_r = data->min_r - x_cp;
 	data->max_r = data->max_r - x_cp;
@@ -23,7 +23,7 @@ static void ft_translation_to_zero(t_data *data, double x_cp, double y_cp)
 	data->max_i = data->max_i - y_cp;
 }
 
-static void ft_scale(t_data *data, double ratio)
+static void ft_scale(t_data *data, long double ratio)
 {
 	data->min_r = data->min_r * ratio;
 	data->max_r = data->max_r * ratio;
@@ -67,18 +67,25 @@ static void	ft_zoom_image(t_data *data, int button, double x, double y)
 
 int	ft_mouse_hook(int button, int x, int y, t_data *data)
 {
-	double	x_cp;
-	double	y_cp;
+	long double	x_cp;
+	long double	y_cp;
 //	ft_printf("button: %d\n", button);
 	if (button == SCROLL_UP || button == SCROLL_DOWN)
 	{
-		x_cp = ((data->max_r - data->min_r) * (double)x) / WIDTH + data->min_r;
-		y_cp = ((data->max_i - data->min_i) * (double)y) / HEIGHT + data->min_i; 
+		x_cp = ((data->max_r - data->min_r) * (long double)x) / (long double)WIDTH + data->min_r;
+		y_cp = ((data->max_i - data->min_i) * (long double)y) / (long double)HEIGHT + data->min_i; 
 		ft_translation_to_zero(data, x_cp, y_cp);
 		if (button == SCROLL_DOWN)
+//		{
+//			if (data->max_iter > 4)
+//				data->max_iter -= 4;
 			ft_scale(data, ZOOM_IN_RATIO);
+//		}
 		else
+//		{
 			ft_scale(data, ZOOM_OUT_RATIO);
+//			data->max_iter += 4;
+//		}
 		ft_translation_back(data, x_cp, y_cp);
 	}
 	return (0);
